# docs/swagger.yaml
basePath: /api
definitions:
...
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  LoggedIn:
    description: LoggedIn represents the candidat details after successful login.
    properties:
      ID:
        description: ID is the unique identifier for the candidat.
        type: string
      email:
        description: Email is the email address of the candidat.
        type: string
      first_name:
        description: Name is the name of the candidat.
        type: string
      last_name:
        description: Name is the name of the candidat.
        type: string
      workCompanyId:
        description: CompanyID is the unique identifier for the candidat company.
        type: string
    type: object
  LoggedInResponse:
    description: LoggedInResponse represents the response structure after successful
      login.
    properties:
      accessToken:
        description: AccessToken is the token obtained after successful login for
          authentication purposes.
        type: string
      user:
        allOf:
        - $ref: '#/definitions/LoggedIn'
        description: User is the structure containing details of the logged-in user.
    type: object
...
  Signin:
    description: Signin represents the information required for signing in candidat.
    properties:
      email:
        description: Email is the email address of the user. It is required, should
          be a valid email, and maximum length is 255 characters.
        maxLength: 255
        type: string
      password:
        description: Password is the user's password. It is required, and its length
          should be between 10 and 255 characters.
        maxLength: 255
        minLength: 10
        type: string
    required:
    - email
    - password
    type: object
  Signup:
    description: Signup represents the account information for signing up.
    properties:
      companyName:
        description: CompanyName is the name of the user's company. It is required
          and should be between 2 and 255 characters.
        maxLength: 255
        minLength: 2
        type: string
      email:
        description: Email is the email address of the user. It is required, should
          be a valid email, and maximum length is 255 characters.
        maxLength: 255
        type: string
      firstName:
        description: Firstname is the first name of the user. It should be between
          3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
      lastName:
        description: Lastname is the last name of the user. It should be between 3
          and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      password:
        description: Password is the user's password. It is required, and its length
          should be between 10 and 255 characters.
        maxLength: 255
        minLength: 10
        type: string
    required:
    - companyName
    - email
    - firstName
    - lastName
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate and log in a user.
      parameters:
      - description: User query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Signin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoggedInResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      summary: Signin
      tags:
      - Authentification
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account.
      parameters:
      - description: User query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Signup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      summary: Signup
      tags:
      - Authentification
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"